apps/api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.20
PyJWT==2.10.1
argon2-cffi==23.1.0
email-validator==2.2.0
httpx==0.28.1
edge-tts==7.2.7
Pillow==10.4.0
pytesseract==0.3.13
pycld2==0.42
pykakasi==2.3.0
pypinyin==0.53.0
